+++ packages/components/select-option/index.ts
@@ -1,0 +1,11 @@
+import { withInstall } from '../utils'
+
+import SelectOption from './select-option.vue'
+
+
+
+
+export const LeSelectOption = withInstall(SelectOption)
+// export default LeSelect
+
+
