+++ packages/components/select/index.ts
@@ -1,11 +1,9 @@
 import { withInstall } from '../utils'
 import Select from './select.vue'
-import SelectOption from './select-option.vue'
+
 
 export const LeSelect = withInstall(Select)
 
 
-export const LeSelectOption = withInstall(SelectOption)
-// export default LeSelect
