--- packages/components/select/inteface.ts
@@ -1,6 +1,0 @@
-
-export interface SelectOptionProxy {
-  value: string | number,
-  label: string
-  disabled: boolean
-}
