+++ packages/components/select/select.scss
@@ -1,73 +1,45 @@
 @use './var.scss' as *;
 
-.select {
-  &-content {
-    font-size: 14px;
-    display: inline-flex;
-    position: relative;
-    border: 1px solid #ccc;
-    box-sizing: border-box;
-    user-select: none;
-    width: 200px;
-    height: 30px;
-    > input {
-      width: 100%;
-      height: 100%;
-      -webkit-appearance: none;
-      border: 1px solid var(--le-accents-3);
-      outline-color: var(--le-accents-6);
-      cursor: pointer;
-    }
-    &-menu {
-      position: absolute;
-      top: 120%;
-      width: 100%;
-      box-sizing: border-box;
-      max-width: 100%;
-      max-height: 300px;
-      border: 1px solid var(--le-accents-6);
-      // background-color: aqua;
-    }
-    &-multiple {
-      position: absolute;
-      left: 0;
-      top: 50%;
-      transform: translateY(-50%);
-      height: 95%;
-      z-index: 2;
-      max-width: 100%;
-      display: flex;
-      > span {
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-        height: inherit;
-        padding: 0 3px;
-        margin-left: 3px;
-        background-color: var(--le-accents-2);
-        border-radius: 5px;
-        > div {
-          display: flex;
-        }
-      }
-    }
-  }
+$select-default-width:150px;
+$select-default-height:30px;
+$select-default-font-size:17px;
 
-  &-option {
-    height: 30px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 0 6px;
-    &:hover:not(.disabled),
-    &__active:not(.disabled) {
-      background-color: aqua;
-    }
-  }
+:root{
+  --select-default-width:#{$select-default-width};
+  --select-default-height:#{$select-default-height};
+  --select-default-font-size:#{$select-default-font-size};
 }
 
-.disabled {
-  cursor: not-allowed;
+.le-select{
+  &__weapper{
 
-  background-color: #ccc;
+    box-sizing: border-box;
+    display: inline-flex;
+    cursor: pointer;
+    border: 1px solid var(--le-accents-2);
+    border-radius: calc(var(--select-default-font-size) * 0.3);
+    overflow: hidden;
+    &--disabled{
+      cursor: not-allowed;
+      background-color: var(--le-accents-3);
+    }
+  }
+
+  &__input{
+    min-width: var(--select-default-width);
+    min-width: 100%;
+    >input{
+      padding: 0;
+      margin: 0;
+      border: none;
+      outline: none;
+      cursor: pointer;
+      line-height: var(--select-default-height);
+      font-size: var(--select-default-font-size);
+      box-shadow: 0 0 1px 1px inset rgba($color: #000000, $alpha: .2);
+    }
+  }
+
+  
+
 }
